fix(core): remove useBundleDeletedToast, move to usePerspective hook
